+++ src/Compute/Compute/Manual/VirtualMachineScaleSetCreateOrUpdateMethod.cs
@@ -274,6 +274,16 @@ namespace Microsoft.Azure.Commands.Compute.Automation
 
             private async Task<ResourceConfig<VirtualMachineScaleSet>> SimpleParameterSetNormalMode()
             {
+                // Temporary message until after the Ignite 2023 release that should remove these outdated image aliases. 
+                if ((_cmdlet.ImageName != "")
+                    &&
+                    (_cmdlet.ImageName == "CentOS" || _cmdlet.ImageName == "Debian" || _cmdlet.ImageName == "RHEL"
+                     || _cmdlet.ImageName == "UbuntuLTS"))
+                {
+                    string ImageOutdatedMessage = "You are using the image " + _cmdlet.ImageName + ", which is outdated and this image name will be removed in October 2023. Please update to a newer versioned image alias as seen here, [Find and use Azure Marketplace VM images with Azure PowerShell](https://learn.microsoft.com/en-us/azure/virtual-machines/windows/cli-ps-findimage#default-images).";
+                    _cmdlet.WriteInformation(ImageOutdatedMessage, new string[] { "PSHOST" });
+                }
+
                 ImageAndOsType = await _client.UpdateImageAndOsTypeAsync(
                         ImageAndOsType, _cmdlet.ResourceGroupName, _cmdlet.ImageName, Location);
                
@@ -452,6 +462,16 @@ namespace Microsoft.Azure.Commands.Compute.Automation
                 checkFlexibleOrchestrationModeParams();
                 int platformFaultDomainCountFlexibleDefault = 1;
                 SwitchParameter singlePlacementGroupFlexibleDefault = false;
+
+                // Temporary message until after the Ignite 2023 release that should remove these outdated image aliases. 
+                if ((_cmdlet.ImageName != "")
+                    &&
+                    (_cmdlet.ImageName == "CentOS" || _cmdlet.ImageName == "Debian" || _cmdlet.ImageName == "RHEL"
+                     || _cmdlet.ImageName == "UbuntuLTS"))
+                {
+                    string ImageOutdatedMessage = "You are using the image " + _cmdlet.ImageName + ", which is outdated and this image name will be removed in October 2023. Please update to a newer versioned image alias as seen here, [Find and use Azure Marketplace VM images with Azure PowerShell](https://learn.microsoft.com/en-us/azure/virtual-machines/windows/cli-ps-findimage#default-images).";
+                    _cmdlet.WriteInformation(ImageOutdatedMessage, new string[] { "PSHOST" });
+                }
 
                 ImageAndOsType = await _client.UpdateImageAndOsTypeAsync(
                         ImageAndOsType, _cmdlet.ResourceGroupName, _cmdlet.ImageName, Location);